fix: improve goroutine safety detection for better test reliability
- Fix amortized sampling to always check first 10 accesses
- Always verify ownership during race conditions on first access
- Remove token multiplication that could cause collisions
- Ensure early detection of concurrent access violations

This fixes test failures while maintaining performance benefits.
