+++ pkg/kernel/kbuffer/BUILD.bazel
@@ -16,7 +16,10 @@ go_test(
     timeout = "short",
     srcs = glob(
         ["*_test.go"],
-        exclude = ["*_bench_test.go"],
+        exclude = [
+            "*_bench_test.go",
+            "*_race_test.go",  # Exclude intentional data race tests
+        ],
     ),
     embed = [":kbuffer"],
     race = "on",
@@ -66,6 +69,16 @@ go_test(
 )
 
 go_test(
+    name = "race_test",
+    size = "small",
+    timeout = "short",
+    srcs = glob(["*_race_test.go"]),
+    embed = [":kbuffer"],
+    race = "off",  # Intentionally disabled for data race testing
+    tags = ["manual"],  # Don't run automatically in CI
+)
+
+go_test(
     name = "kbuffer_test",
     srcs = [
         "global_test.go",
@@ -74,7 +87,9 @@ go_test(
         "kbuffer_bench_test.go",
         "safe_buffer_test.go",
         "safe_sharded_test.go",
+        "unsafe_buffer_race_test.go",
         "unsafe_buffer_test.go",
+        "unsafe_sharded_race_test.go",
         "unsafe_sharded_test.go",
     ],
     embed = [":kbuffer"],
