+++ pkg/kernel/kbuffer/unsafe_sharded_test.go
@@ -2,8 +2,6 @@ package kbuffer
 
 import (
 	"bytes"
-	"runtime"
-	"sync"
 	"testing"
 	"time"
 )
@@ -704,41 +702,8 @@ func TestUnsafeShardedBufferPerformance(t *testing.T) {
 	t.Logf("Operations per second: %.0f", opsPerSec)
 }
 
-// TestUnsafeShardedBufferDataRace intentionally tests for data races.
-func TestUnsafeShardedBufferDataRace(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping race condition test in short mode")
-	}
-
-	// Disable safety checks to allow concurrent access (for testing only!)
-	oldDebugMode := testingSkipSafetyCheck
-	testingSkipSafetyCheck = true // true = skip checks
-	defer func() { testingSkipSafetyCheck = oldDebugMode }()
-
-	buf := newUnsafeShardedBuffer(10000, 8).(*unsafeShardedBuffer)
-
-	// WARNING: This deliberately creates a data race
-	var wg sync.WaitGroup
-	const goroutines = 10
-	wg.Add(goroutines)
-
-	for i := 0; i < goroutines; i++ {
-		go func(id int) {
-			defer wg.Done()
-			data := []byte{byte(id)}
-			for j := 0; j < 100; j++ {
-				// This will corrupt data or panic
-				_, _ = buf.Write(data)
-				runtime.Gosched()
-			}
-		}(i)
-	}
-
-	wg.Wait()
-
-	// Buffer is likely corrupted
-	t.Logf("Data race test completed (buffer likely corrupted): Len=%d", buf.Len())
-}
+// TestUnsafeShardedBufferDataRace is moved to a separate file with !race build tag
+// to exclude it when the race detector is enabled
 
 // TestUnsafeShardedBufferSelectShard tests shard selection logic.
 func TestUnsafeShardedBufferSelectShard(t *testing.T) {
